DOES NOT WORK but its better than before
